Add usg_id and passport_id input fields
We also now need to allow a person to be entered without a unique PIDM. We have
some trouble in that we need to enforce a uniqueness constraint on existing
PIDMs, but not for those that don't have PIDMs. To work around this, for now, we
prefix a unique value for empty PIDM colums with 'ignore-this-value-' to
indicate to users that the PIDM is to be ignored. We need to find a better way
of solving this problem, but I don't have one at the moment.